+++ tests/testthat.R
@@ -1,0 +1,5 @@
+# This file is part of the standard testthat setup
+library(testthat)
+library(sumExtras)
+
+test_check("sumExtras")